requirements.txt
langchain_core
langchain_community
langchain
python-dotenv
streamlit
groq
langchain-groq
PyPDF2